Pull functions out to individual constructs and add WIP readme
