+++ index.ts
@@ -1,3 +1,5 @@
-import { PrerenderLambdaStack } from "./lib/lambda-stack";
+import { PrerenderLambdaStack } from "./lib/prerender-lambda-stack";
+import { PrerenderFunction } from "./lib/prerender-construct";
+import { PrerenderCheckFunction } from "./lib/prerender-check-construct";
 
-export { PrerenderLambdaStack };
+export { PrerenderLambdaStack, PrerenderFunction, PrerenderCheckFunction };
